% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affil_df.R
\name{affil_df}
\alias{affil_df}
\alias{affil_data}
\title{Search Affiliation Content on SCOPUS}
\usage{
affil_df(affil_id, verbose = TRUE, ...)

affil_data(affil_id, verbose = TRUE, ...)
}
\arguments{
\item{affil_id}{Affiliation ID number.}

\item{verbose}{Print diagnostic messages}

\item{...}{Arguments to be passed to \code{\link{author_search}}}
}
\value{
List of entries from SCOPUS
}
\description{
Searches SCOPUS to get information about documents on an affiliation
}
\note{
The \code{affil_data} command will return the list of all entries as well as
the \code{data.frame}
}
\seealso{
\code{\link{get_author_info}}
}
