% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affil_list_to_df.R
\name{affil_list_to_df}
\alias{affil_list_to_df}
\title{List of SCOPUS Entries to List of Affiliations Data Frames}
\usage{
affil_list_to_df(affils)
}
\arguments{
\item{affils}{List of affiliations, from \code{\link{entries_to_affil_list}}}
}
\value{
A \code{data.frame} of affiliation information.  A column named
\code{"index"} denotes the element of the object \code{affils} that the row
corresponds to
}
\description{
Take a SCOPUS entry and transform it to a data frame of
affiliations
}
