% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_possible_affils.R
\name{all_possible_affils}
\alias{all_possible_affils}
\title{Find all affiliations}
\usage{
all_possible_affils(entries)
}
\arguments{
\item{entries}{list of entries from SCOPUS,
usually from \code{\link{author_search}} result}
}
\value{
A \code{data.frame} of affiliations
}
\description{
Take a SCOPUS entry and transform it to a data frame of
all affiliations listed in there
}
