% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/author_search.R
\name{author_search}
\alias{author_search}
\title{Search Author Content on SCOPUS}
\usage{
author_search(au_id, api_key = NULL,
  http = "http://api.elsevier.com/content/search/scopus", count = 100,
  verbose = TRUE, ...)
}
\arguments{
\item{au_id}{Author ID number}

\item{api_key}{API Key for Elsevier}

\item{http}{Address for scopus api}

\item{count}{number of records to retrieve (below 100)}

\item{verbose}{Print diagnostic messages}

\item{...}{Arguments to be passed to \code{\link{GET}}}
}
\value{
List of entries from SCOPUS
}
\description{
Searches SCOPUS to get information about documents on an author.
}
\seealso{
\code{\link{get_author_info}}
}

