% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_search.R
\name{author_search_by_affil}
\alias{author_search_by_affil}
\title{Search Authors by Affiliation on SCOPUS}
\usage{
author_search_by_affil(affil_id, searcher = "AF-ID", ...)
}
\arguments{
\item{affil_id}{Affiliation ID number}

\item{searcher}{Identifier for Affiliation ID.  Do not change unless you
know exactly what the API calls for.}

\item{...}{Arguments to be passed to \code{\link{GET}}}
}
\value{
List of entries from SCOPUS
}
\description{
Searches SCOPUS to get information about authors with
a certain affiliation
}
\seealso{
\code{\link{get_author_info}}
}
