% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_retrieval.R
\name{citation_retrieval}
\alias{citation_retrieval}
\title{SCOPUS Citation Retrieval}
\usage{
citation_retrieval(scopus_id = NULL, pii = NULL, doi = NULL,
  pubmed_id = NULL, ...)
}
\arguments{
\item{scopus_id}{Scopus Identifier}

\item{pii}{Scopus Identifier}

\item{doi}{Scopus Identifier}

\item{pubmed_id}{Scopus Identifier}

\item{...}{Arguments to be passed to \code{\link{generic_elsevier_api}}}
}
\value{
List of elements, similar to \code{\link{generic_elsevier_api}}
}
\description{
SCOPUS Citation Retrieval
}
\examples{
api_key = get_api_key(NULL, error = FALSE)
if (!is.null(api_key)){
   x = citation_retrieval(pii = "S1053811915002700",
   verbose = FALSE)
}
}
\seealso{
\code{\link{generic_elsevier_api}}
}
