% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries_to_affil_list.R
\name{collapse_affil}
\alias{collapse_affil}
\title{Collapse Affiliations to one row}
\usage{
collapse_affil(affils, collapse = ";")
}
\arguments{
\item{affils}{Data frame of affiliations to be collapsed
usually from \code{\link{author_search}}}

\item{collapse}{What should values be collapsed using as a separator}
}
\value{
A \code{data.frame} of affiliation information
}
\description{
Take an individual SCOPUS entry and transform it to a data frame of
affiliations
}

