% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_author_info.R
\name{complete_multi_author_info}
\alias{complete_multi_author_info}
\title{Get Complete Author Information and ID from Scopus}
\usage{
complete_multi_author_info(au_id = NULL, api_key = NULL, verbose = TRUE, ...)
}
\arguments{
\item{au_id}{vector of Author IDs}

\item{api_key}{Elsevier API key}

\item{verbose}{Print messages from specification}

\item{...}{options to pass to \code{\link{generic_elsevier_api}}}
}
\value{
List of information
}
\description{
Get Complete Author Information and ID from Scopus
}
