% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsevier_authenticate.R
\name{elsevier_authenticate}
\alias{elsevier_authenticate}
\title{Authenticate API Key and get Token}
\usage{
elsevier_authenticate(api_key = NULL, api_key_error = TRUE,
  verbose = TRUE)
}
\arguments{
\item{api_key}{Elsevier API key}

\item{api_key_error}{Should there be an error if no API key?}

\item{verbose}{Print messages from specification}
}
\value{
List of content, the \code{GET} request,
and the token
}
\description{
Authenticate API Key and get Token
}
\examples{
if (have_api_key()) {
   auth = elsevier_authenticate()
}
}
