% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries_to_affil_list.R
\name{entries_to_affil_list}
\alias{entries_to_affil_list}
\title{List of SCOPUS Entries to List of Affilations Data Frames}
\usage{
entries_to_affil_list(entries)
}
\arguments{
\item{entries}{list of entries from SCOPUS,
usually from \code{\link{author_search}} result}
}
\value{
A \code{data.frame} of affiliation information
}
\description{
Take a SCOPUS entry and transform it to a data frame of
affiliations
}

