% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries_to_citation_df.R
\name{entries_to_citation_df}
\alias{entries_to_citation_df}
\title{List of SCOPUS Entries to Data Frame of Citations}
\usage{
entries_to_citation_df(entries)
}
\arguments{
\item{entries}{list of entries from SCOPUS,
usually from \code{\link{author_search}} result}
}
\value{
A \code{data.frame} of citation information
}
\description{
Take a SCOPUS entry list and transform it to a data frame of
citation and article information
}
