% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries_to_df.R
\name{entries_to_df}
\alias{entries_to_df}
\alias{entries_to_df2}
\title{Convert Entries into a data.frame}
\usage{
entries_to_df(entries, au_id = NULL, verbose = TRUE)

entries_to_df2(entries, verbose = TRUE)
}
\arguments{
\item{entries}{Entries from the output of a command}

\item{au_id}{Author ID number. Overrides any first/last name argument}

\item{verbose}{Print diagnostic messages}
}
\value{
Data frame of records
}
\description{
Converts a list of entries into a \code{data.frame} of records
}
