% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries_to_affil_list.R
\name{entry_to_affil}
\alias{entry_to_affil}
\title{SCOPUS Entry to Affiliation}
\usage{
entry_to_affil(x, all_affils)
}
\arguments{
\item{x}{individual entry from SCOPUS,
usually from \code{\link{author_search}}}

\item{all_affils}{Affiliation data.frame from \code{\link{all_possible_affils}}}
}
\value{
A \code{data.frame} of affiliation information
}
\description{
Take an individual SCOPUS entry and transform it to a data frame of
affiliations
}
