% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_affiliation_info.R
\name{get_affiliation_info}
\alias{get_affiliation_info}
\title{Get Affiliation Information and ID from Scopus}
\usage{
get_affiliation_info(
  affil_id = NULL,
  affil_name = NULL,
  api_key = NULL,
  verbose = FALSE,
  headers = NULL
)
}
\arguments{
\item{affil_id}{ID of affiliation}

\item{affil_name}{name of affiliation}

\item{api_key}{Elsevier API key}

\item{verbose}{Print messages from specification}

\item{headers}{Headers passed to \code{\link[httr]{add_headers}},
passed to \code{\link[httr]{GET}}}
}
\value{
A \code{data.frame} of affiliation information
}
\description{
Uses SCOPUS affiliation search to identify affiliation identification
information
}
