% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_coauthors.R
\name{get_all_coauthors}
\alias{get_all_coauthors}
\title{Get all Co-Authors}
\usage{
get_all_coauthors(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{author_df}}}
}
\value{
A list of output, including a vector of author_ids
}
\description{
Get all Co-Authors
}
\examples{
 \dontrun{
 get_all_coauthors(last_name = "muschelli")
 }
}
