% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_api_key.R
\name{get_api_key}
\alias{get_api_key}
\title{Find API Key for Elsevier}
\usage{
get_api_key(api_key = NULL)
}
\arguments{
\item{api_key}{Elsvier API key}
}
\value{
API key
}
\description{
Determines if \code{option(elsevier_api_key)} is set.
If not, it stops and returns an error.  If so, returns the value.
}

