% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_author_info.R
\name{get_author_info}
\alias{get_author_info}
\title{Get Author Information and ID from Scopus}
\usage{
get_author_info(last_name, first_name = NULL, api_key,
  http = "http://api.elsevier.com/content/search/author")
}
\arguments{
\item{last_name}{last name of author}

\item{first_name}{first name of author}

\item{api_key}{Elsvier API key}

\item{http}{Author API http}
}
\value{
Matrix of data
}
\description{
Uses SCOPUS author search to identify author identification
information for
}

