% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next.R
\name{get_links}
\alias{get_links}
\alias{next_url}
\alias{get_next}
\title{Get Links for next/first/last query}
\usage{
get_links(result)

next_url(result)

get_next(result, ...)
}
\arguments{
\item{result}{Object (list) with an element named \code{content}, usually
from \code{generic_elsevier_api}}

\item{...}{Options passed to \code{\link{GET}}}
}
\value{
A \code{data.frame} or a vector of characters
}
\description{
Get Links for next/first/last query
}
\examples{
\dontrun{
result <- generic_elsevier_api(
query = "ISSN(0004-3702) AND YEAR(2001)",
search_type = "scopus")
next_result = get_next(result)

}

}
