% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_key.R
\name{inst_token_header}
\alias{inst_token_header}
\alias{auth_token_header}
\title{Add Institution or Authorization Token}
\usage{
inst_token_header(token)

auth_token_header(token)
}
\arguments{
\item{token}{Elsevier API token, usually from
\code{\link{elsevier_authenticate}}}
}
\value{
An object of class \code{request}
}
\description{
Add Institution or Authorization Token
}
