% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_author_info.R
\name{multi_author_info}
\alias{multi_author_info}
\title{Get Relevant Authors Information from IDs from Scopus}
\usage{
multi_author_info(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{get_complete_author_info}}}
}
\value{
Data.frame of information
}
\description{
Get Relevant Authors Information from IDs from Scopus
}
\examples{
\dontrun{
multi_author_info(au_id =  "22968535800", "40462056100"))
}
}
\seealso{
\code{\link{get_complete_author_info}}
}
