% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_possible_affils.R
\name{nonull}
\alias{nonull}
\title{Remove NULL}
\usage{
nonull(x, replace = NA)
}
\arguments{
\item{x}{Vector from a list}

\item{replace}{Value to replace NULL with}
}
\value{
Vector
}
\description{
Removes NULL values from a vector from a list
}

