% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.scopus_api_key.R
\name{print.scopus_api_key}
\alias{print.scopus_api_key}
\title{Print method for Scopus API key}
\usage{
\method{print}{scopus_api_key}(x, reveal = FALSE, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{reveal}{Should the API key be revealed}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print method for Scopus API key
}
\examples{
x = "asdf"
class(x) = "scopus_api_key"
print(x)
print(x, reveal = TRUE)
}
