% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_affil_name.R
\name{process_affiliation_name}
\alias{process_affiliation_name}
\title{Process Affiliation Name}
\usage{
process_affiliation_name(
  affil_id = NULL,
  affil_name = NULL,
  api_key = NULL,
  verbose = TRUE
)
}
\arguments{
\item{affil_id}{Affiliation ID number.}

\item{affil_name}{affiliation name}

\item{api_key}{Elsevier API key}

\item{verbose}{Print diagnostic messages}
}
\value{
List of affiliation name and affiliation ID
}
\description{
Process Affiliation ID and names for generic use
}
