% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_df.R
\name{process_author_name}
\alias{process_author_name}
\title{Process Author Name}
\usage{
process_author_name(au_id, last_name, first_name, api_key = NULL,
  verbose = TRUE)
}
\arguments{
\item{au_id}{Author ID number. Overrides any first/last name argument}

\item{last_name}{last name of author}

\item{first_name}{first name of author}

\item{api_key}{Elsvier API key}

\item{verbose}{Print diagnostic messages}
}
\value{
List of first/last name and author ID
}
\description{
Process author ID and names for generic use
}
\note{
This function is really to avoid duplication
}
