% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cto.R
\name{read_cto}
\alias{read_cto}
\alias{read_cto_long}
\title{Read Citation Overview (CTO) File}
\usage{
read_cto(file)

read_cto_long(file)
}
\arguments{
\item{file}{CSV of CTO export from Scopus}
}
\value{
A list of the data, year columns, and
header information
}
\description{
Read Citation Overview (CTO) File
}
\examples{
file = system.file("extdata", "CTOExport.csv", package = "rscopus")
citations_over_time = read_cto(file)
citations_over_time = citations_over_time$data
}
