% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_search.R
\name{scopus_search}
\alias{scopus_search}
\alias{sciencedirect_search}
\alias{scidir_search}
\title{SCOPUS Search}
\usage{
scopus_search(query, api_key = NULL, count = 200,
  view = c("STANDARD", "COMPLETE"), start = 0, verbose = TRUE,
  max_count = 20000,
  http = "https://api.elsevier.com/content/search/scopus",
  headers = NULL, wait_time = 0, ...)

sciencedirect_search(count = 100, ...)

scidir_search(count = 100, ...)
}
\arguments{
\item{query}{Query string to search on SCOPUS}

\item{api_key}{API Key for Elsevier}

\item{count}{number of records to retrieve (below 200 for STANDARD,
below 25 for COMPLETE views, see
\url{https://dev.elsevier.com/api_key_settings.html}).}

\item{view}{type of view to give, see
\url{https://api.elsevier.com/documentation/ScopusSearchAPI.wadl}}

\item{start}{where should the records start gathering}

\item{verbose}{Print diagnostic messages}

\item{max_count}{Maximum count of records to be returned.}

\item{http}{Address for scopus API}

\item{headers}{additional headers to be added to
\code{\link{add_headers}}}

\item{wait_time}{The time in seconds to wait across consecutive
requests of a single search (when records > 25)}

\item{...}{Arguments to be passed to the query list for
\code{\link{GET}}}
}
\value{
List of entries from SCOPUS
}
\description{
This function wraps \code{\link{generic_elsevier_api}} to give a
scopus search from the Elsevier Scopus Search API

Searches SCOPUS to get information about documents on an author.
}
\examples{
if (have_api_key()) {
res = scopus_search(query = "all(gene)", max_count = 20,
count = 10)
df = gen_entries_to_df(res$entries)
head(df$df)
sci_res = sciencedirect_search(query = "heart+attack AND text(liver)",
max_count = 30, count = 25)
sci_df = gen_entries_to_df(sci_res$entries)
Sys.sleep(2)
nt = sciencedirect_search(query = "title(neurotoxin)", max_count = 20,
count = 10, wait_time = 1)
nt_df = gen_entries_to_df(nt$entries)
nt_df = nt_df$df
}
}
