% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API Key for Elsevier}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{Elsvier API key}
}
\description{
Sets Elsevier API key using
if \code{option(elsevier_api_key)}
}

