% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/affil_search.R
\name{affil_search}
\alias{affil_search}
\title{Search Author Content on SCOPUS}
\usage{
affil_search(affil_id, api_key = NULL,
  http = "http://api.elsevier.com/content/search/scopus", count = 100,
  verbose = TRUE, ...)
}
\arguments{
\item{api_key}{API Key for Elsevier}

\item{http}{Address for scopus api}

\item{count}{number of records to retrieve (below 100)}

\item{verbose}{Print diagnostic messages}

\item{...}{Arguments to be passed to \code{\link{GET}}}

\item{au_id}{Affiliation ID number}
}
\value{
List of entries from SCOPUS
}
\description{
Searches SCOPUS to get information about documents on an author.
}
\seealso{
\code{\link{get_author_info}}
}

