% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/affil_search.R
\name{affil_search}
\alias{affil_search}
\title{Search Author Content on SCOPUS}
\usage{
affil_search(affil_id, searcher = "AF-ID", ...)
}
\arguments{
\item{affil_id}{Affiliation ID number}

\item{searcher}{Identifer for Affiliation ID.  Do not change unless you
know exactly what the API calls for.}

\item{...}{Arguments to be passed to \code{\link{GET}}}
}
\value{
List of entries from SCOPUS
}
\description{
Searches SCOPUS to get information about documents on an author.
}
\seealso{
\code{\link{get_author_info}}
}

