% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/author_df.R
\name{author_df}
\alias{author_df}
\title{Search Author Content on SCOPUS}
\usage{
author_df(au_id, last_name, first_name, api_key = NULL, verbose = TRUE, ...)
}
\arguments{
\item{au_id}{Author ID number. Overrides any first/last name argument}

\item{last_name}{last name of author}

\item{first_name}{first name of author}

\item{api_key}{Elsvier API key}

\item{verbose}{Print diagnostic messages}

\item{...}{Arguments to be passed to \code{\link{author_search}}}
}
\value{
List of entries from SCOPUS
}
\description{
Searches SCOPUS to get information about documents on an author.
}
\examples{
\dontrun{
author_df(last_name = "Caffo", first_name = "Brian")
}
}
\seealso{
\code{\link{get_author_info}}
}

