% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_search.R
\name{author_search}
\alias{author_search}
\title{Search Author Content on SCOPUS}
\usage{
author_search(au_id, api_key = NULL,
  http = "http://api.elsevier.com/content/search/scopus", count = 25,
  verbose = TRUE, facets = "subjarea(sort=fd)", searcher = "AU-ID",
  max_count = Inf, ...)
}
\arguments{
\item{au_id}{Author ID number}

\item{api_key}{API Key for Elsevier}

\item{http}{Address for scopus api}

\item{count}{number of records to retrieve (below 25, see
\url{http://dev.elsevier.com/api_key_settings.html})}

\item{verbose}{Print diagnostic messages}

\item{facets}{Facets sent in query.  See \url{http://dev.elsevier.com/api_docs.html}}

\item{searcher}{Identifier for author ID.  Do not change unless you
know exactly what the API calls for.}

\item{max_count}{Maximum count of records to be returned.}

\item{...}{Arguments to be passed to the query list for
\code{\link{GET}}}
}
\value{
List of entries from SCOPUS
}
\description{
Searches SCOPUS to get information about documents on an author.
}
\seealso{
\code{\link{get_author_info}}
}
