% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_entries_to_df.R
\name{gen_entries_to_df}
\alias{gen_entries_to_df}
\title{Generally Convert Entries into a list of \code{data.frame}s}
\usage{
gen_entries_to_df(entries)
}
\arguments{
\item{entries}{Entries from the output of a command}
}
\value{
List of \code{data.frame}s from entries
}
\description{
Generally Convert Entries into a list of \code{data.frame}s
}
