% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_affiliation_info.R
\name{get_affiliation_info}
\alias{get_affiliation_info}
\title{Get Affiliation Information and ID from Scopus}
\usage{
get_affiliation_info(affil_id = NULL, affil_name = NULL,
  api_key = NULL, verbose = FALSE)
}
\arguments{
\item{affil_id}{ID of affiliation}

\item{affil_name}{name of affiliation}

\item{api_key}{Elsevier API key}

\item{verbose}{Print messages from specification}
}
\value{
A \code{data.frame} of affiliation information
}
\description{
Uses SCOPUS affiliation search to identify affiliation identification
information
}
