% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_author_info.R
\name{get_author_info}
\alias{get_author_info}
\title{Get Relevant Author Information and ID from Scopus
in Data Frame}
\usage{
get_author_info(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{get_complete_author_info}}}
}
\value{
Data.frame of information
}
\description{
Uses SCOPUS author search to identify author identification
information in a workable format
}
\seealso{
\code{\link{get_complete_author_info}}
}
