% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_author_info.R
\name{get_complete_author_info}
\alias{get_complete_author_info}
\title{Get Complete Author Information and ID from Scopus}
\usage{
get_complete_author_info(
  last_name = NULL,
  first_name = NULL,
  affil_id = NULL,
  affil_name = NULL,
  api_key = NULL,
  http = "https://api.elsevier.com/content/search/author",
  query = NULL,
  count = 200,
  start = 0,
  verbose = TRUE,
  au_id = NULL,
  headers = NULL,
  ...
)
}
\arguments{
\item{last_name}{last name of author}

\item{first_name}{first name of author}

\item{affil_id}{ID of affiliation (optional)}

\item{affil_name}{name of affiliation}

\item{api_key}{Elsevier API key}

\item{http}{Author API http}

\item{query}{Additional query info, added using \code{+AND+} to original query}

\item{count}{maximum number of records to retrieve}

\item{start}{index to start on.  Only necessary if a large number of
records retrieved}

\item{verbose}{Print messages from specification}

\item{au_id}{Author ID number, will override first/last combination if
specified}

\item{headers}{Headers passed to \code{\link[httr]{add_headers}},
passed to \code{\link[httr]{GET}}}

\item{...}{options to pass to \code{\link[httr]{GET}}}
}
\value{
List of information
}
\description{
Uses SCOPUS author search to identify author identification
information
}
\examples{
if (is_elsevier_authorized()) {
res = get_complete_author_info(
last_name = "Muschelli",
first_name = "John",
verbose = FALSE)
}
}
