% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_non_ascii.R
\name{replace_non_ascii}
\alias{replace_non_ascii}
\title{Replace non-ASCII characters}
\usage{
replace_non_ascii(string)
}
\arguments{
\item{string}{Character vector of values to be replaced}
}
\value{
Character vector
}
\description{
Replaces non-ASCII characters from a last or first name
}
\seealso{
Taken from \url{https://stackoverflow.com/questions/20495598/replace-accented-characters-in-r-with-non-accented-counterpart-utf-8-encoding}
}
